requirements.txt
flask
Flask-SocketIO
eventlet
google-generativeai
python-dotenv
chromadb
sentence-transformers
PyMuPDF
markdown2
